Initialize ed25519::Keyring and pass it through to client Sessions
...also a "kms keygen" command!

This processes the ed25519-dalek provider-related portions of the config
file and uses them to initialize `Box<Signer>` objects and create an
`ed25519::Keyring` of them.

The `Signer` objects are now marked as `Send` safe as of Signatory 0.5.1
and the keyring is marked as `RefUnwindSafe`, allowing `Arc`s of the
keyring to pass through the `catch_unwind` barrier which `Client` uses
to restart failing sessions.

The result is `Session` structs now hold a copy of the `Arc<Keyring>`
and can use it to produce signatures... at least as soon as we add a
"sign" command.