Make `ed25519::KeyRing` a global static
Changes the Ed25519 signer keyring from an `Arc<Mutex<_>>` to a static
global `RwLock<KeyRing>`.

This avoids having to move `KeyRing` across the unwind boundary, which
means we no longer have to `impl RefUnwindSafe` which is a big plus.

It also means `KeyRing` is accessible from any scope and we don't have
to worry about passing it around as an argument.

This additionally puts the `SecretConnection` key management more or
less back to where it was in the original PR, factored out of `KeyRing`
and back into `RunCommand`. There and back again.