Add 'yubihsm test' command
This performs an Ed25519 stress test on a YubiHSM, signing messages as
quickly as it can.

If an error occurs during signing, it sleeps briefly and then resumes
the stress test, ideally picking right back up again once the problem
has been resolved.

This code should be panic free (in a perfect world). If it panics,
there's a bug. So far in my initial tests with the USB driver, it
has survived repeatedly disconnecting and reconnecting the YubiHSM2 as
well as moving it to different USB ports.