Re-enable deny(warnings); make a lib crate
Now that #55 is landed, we can re-enable `#![deny(warnings)]` which is a
best practice.

This commit does that, and peppers remaining unused code with
`#![allow(dead_code)]`.

Additionally, to ensure `#![deny(missing_docs)]` is enforced and to
follow (evolving) Abscissa conventions, this refactors the app as a
library with a small `main.rs` stub.