Adds validator integration testing capability
This commit adds the necessary componentry to be able to run the
integration testing harness.

Two important scripts are included to help in (1) generating
configuration for the new Docker image, and (2) running the test
harness.

The Dockerfile is updated to a multi-stage build that grabs the remote
validator test harness binary from the remote validator image. This new
Docker image will also necessarily contain the relevant configuration to
run the harness and KMS.

The CircleCI configuration is updated to include the new Docker image.