setup: Add whitespace tolerance reading admin phrase; remove rustyline
I added rustyline to allow "correctability" when reading the 24-word
admin mnemonic phrase from stdin. But it seems like a very heavy, scary
dependency for such sensitive data (possibly logging it to history?).

This change removes it, uses io::stdin().read_line(), but strips
whitespace, which inadvertently allows the password to be "editable"
without the need for rustyline.