yubihsm: Support for exporting/importing wrapped (encrypted) keys
Adds support directly to the KMS to make encrypted backups of keys which
have been generated inside of the YubiHSM (retroactively, as there
exists a similar feature as part of key generation).

Also adds a corresponding command to import the encrypted backups.
There was already a `tmkms yubihsm keys import` command so this
overloads the command to handle both importing existing
`priv_validator.json` files (the previous functionality) and importing
encrypted backups, with the former ideally going away with time.