last_sign_state: Atomic writes
Uses the `atomicwrites` crate to ensure that the validator state JSON
files are serialized as a consistent snapshots.

This eliminates the potential for filesystem races where e.g. two copies
of the state are written to the same file, rendering its contents
invalid (as was seen in earlier test failures).