Track chain state in registry (fixes #60)
This commit moves the last sign state tracking for chains into the
global chain registry.

It also allows configurable locations for the chain state tracking
files, which should probably make it easier to ensure they don't clobber
each other in tests.

This doesn't fully implement the double signing plan in #60 but at this
point I think we're close enough. The remaining item is (optionally)
running a user-specified command on startup to query the current block
height.