tendermint-rs: Initial JSONRPC over HTTP client
Adds a hyper-based HTTP client for making JSONRPC requests.

Uses a slightly old version of hyper (0.10) which does not yet depend on
futures/tokio, the idea being updating to a fully async version should
be possible in the future.

Adds an integration test which can be used against gaiad locally. This
isn't run in CI (yet) but it is built in CI at least to ensure changes
still allow it to compile.