tendermint-rs: Add tests for `/block_results` RPC endpoint
Previously there was only a live integration test.

This tests against a vendored RPC result for cosmoshub-1, and does a
number of assertions on how the data is parsed.

It also simplifies the API somewhat through the addition of some custom
parsers that eliminate superfluous `Option` types (e.g. where an empty
`Vec` would suffice).