Send RemoteSignerError response on double sign (closes #249)
Previously double signing would abort the connection.

However, there is a semi-valid use case for returning an error message
instead: when concurrent validators on the same chain are sending
signing messages. This was proposed by @mdyring in #249.

Ideally there should be coordination (i.e. between KMS instances) as to
which validator is currently active, as this approach depends critically
on the KMS's double signing prevention and encourages configurations
where multiple validator instances are attempting to sign
simultaneously. This runs the risk that a bug in the KMS's double
signing detection could be singularly responsible for a double sign
event.

However, without something like this, it isn't possible for the KMS to
service two validators simultaneously, so this seems like an OK start.