yubihsm-server: Allow CLI commands to use loopback connection
This provides configuration for allowing `tmkms yubihsm` CLI commands to
connect to a running tmkms process via the `connector_server.laddr`
address.

This provides a way to administer the KMS while it's running without
modifying the configuration file.

Additionally, it allows providing an auth key override, allowing the
administrator to input a password to perform commands (ala "sudo").