tendermint-rs: Add TendermintConfig and Error(Kind) types
Adds types which describes the structure of `config.toml` files, and
support for parsing them with serde and the `toml` crate.

This should describe every configuration field generated by the default
boilerplate.

Additionally this commit changes the `Error` type in `tendermint-rs`
from an enum to a struct, renaming the original `Error` enum to
`ErrorKind`. This allows incorporating additional information into each
error: this commit uses it to add a "msg" field to each error which is
needed to get reasonable parse errors from config files, however since
it's built on the `failure` crate and using a `failure::Context`, this
will also capture a backtrace at the time the error is generated.

Last but certainly not least, this adds parsers and serializers for the
`node_key.json` and `priv_validator_key.json` configuration files as
well, along with a new `tendermint::private_key::PrivateKey` type for
Ed25519 keypairs.