yubihsm/setup: Test vectors for key derivation hierarchy
I am planning to address one of the findings (low severity) of a recent
security audit, as well and replace the key derivation implementation
with one I've extracted into an external crate for reuse:

https://crates.io/crates/hkd32

However, the key derivation performed by `tmkms yubihsm setup` does not
presently have any tests (it was hastily written in order to meet the
deadlines for the Cosmos Hub launch).

This PR adds test vectors which capture the current state of the
implementation, so as to ensure that followup PRs derive the same key
hierarchy.