yubihsm setup: Collect 256-bits entropy from both RNGs
Previously it collected 128-bits entropy from each RNG for a total of
256-bits.

This commit expands it to collect 256-bits entropy from each RNG for a
total of 512-bits of input key material.

This ensures that the desired amount of 256-bits is collected even in
the event that one of the two RNGs fails silently.