softsign: Add subcommand; move keygen under it
The `tmkms keygen` command is exclusively for the softsign backend,
but that isn't immediately clear.

This commit adds a `tmkms softsign` subcommand, which is gated on Cargo
features, and moves the keygen command to `tmkms softsign keygen`.

This also makes room for additional subcommands (I plan on adding a
`tmkms softsign import` command to do key type conversion).