softsign: Fix private key decoding + import command (closes #290)
This fixes the inconsistencies between Base64 and binary/"raw" private
keys, and adds key conversion which can take a `priv_validator.json`
file or a "raw" (i.e. binary) private key and convert it to the Base64
format used by the `softsign` backend.

Additionally, it adds a `key_format` configuration option allowing it
to load private keys from any of the supported key formats.

This should be sufficient to resolve the problems in #290, although I
also plan on making the private key format a configuration option.