tendermint-rs: Fix feature gating on signatory-dalek + CI tests
There are now several functions which rely on the ability to compute
Ed25519 public keys from private scalars, and use `signatory-dalek` to
do that.

However, those weren't properly `cfg` gated, so in those cases the crate
wouldn't compile.

Additionally the `config` feature didn't pull in `zeroize`, a required
dependency. That was corrected by adding it.