Log block ID and [dup] blocks
Logs a truncated ID of every block that is requested to be signed.

The idea is to help users of the KMS understand what is happening when
multiple concurrent validators attempt to (re)sign the same block.

Additionally this adds a small `[dup]` flag when resiging the same block
is requested, to let users know that the KMS is aware it resigned a
block at the same height/round/step, but that's OK because it has the
same block ID and will produce the same signature, so rather than it
being a "double signing" event, it's merely recomputing the same
signature on the block it originally signed.