Move Amino type prefixes over from KMS
I'd like to compute some Amino type prefixes in something other than the
KMS (namely an oracle feeder) and unfortunately the code to do so is
presently vendored into the KMS, and used to compute these prefixes
(with a TODO to move it elsewhere):

https://github.com/tendermint/kms/blob/master/src/rpc.rs#L45

It seems like these should go with the Amino types. Perhaps the
`compute_prefix` method should be moved into the `amino_rs` crate.