Turn `lite::types::Requester` into an async trait
Requester not being an async trait forced implementations
to block on RPC queries, eg. by spawning a new executor,
which is undesirable considering that we eventually want
the whole codebase to play nice with async/.await,
and not unnecessarily block the enclosing task.

Because trait cannot currently define async functions, this commit
introduces a dependency on the `async-trait` crate, which turns
async fns within traits and their implementations into plain
functions returning `Pin<Box<dyn Future<Output = T + ...>>>`.

The obvious drawback of this approach is that these async fns now
allocate memory on the heap. Hopefully, the language will soon
stabilize the feature, which will help us get rid of these
undesirable heap allocations.

On top of that, the `lite::verifier::verify_bisection_inner` fn had
to be manually adapted to return a boxed future because recursive
async fn are not currently supported in Rust.

An possible alternative would be to rewrite `verify_bisection_inner`
to not be recursive anymore, but that is outside of the scope of
this commit.