Bootstrap rpc proxy for light-node API
Introducing a proxy which serves the Tendermint RPC API with the help of
connect rpc client. Later to be extended to add light client
functionality. This is in nature equivalent to the Go proxy[0].

Instead of reaching for hyper this change introduces warp to build an
http server. This decision should be evaluated very carefully, while
warp is a thinly wrapper around hyper, it brings a lot of very important
tooling (including websockets), it also has its own set of dependencies.

Also currently unclear if the module that implements the rpc API with
lite client functionality should live in the rpc source tree.

Part of #219

[0] https://github.com/tendermint/tendermint/tree/master/light/proxy