light-client: turn Handle into a trait (#401)
* light-client: turn Handle into a trait

As we start to depend on the surface of the `Handle` we benefit from it
being a trait that can be implemented on a per need basis. This will
result in less overhead constructing object graphs in places where we
wannt to assert behaviour of other types in remote places, i.e. the
light-node rpc server. Overall we hope for an increased ease in writing
tests on module level.

Ref #219

* Add changes entry

* Apply suggestions from code review

Co-authored-by: Ismail Khoffi <Ismail.Khoffi@gmail.com>

Co-authored-by: Ismail Khoffi <Ismail.Khoffi@gmail.com>