Rework and generalize the PeerList for easier testing and safer API
This commit changes the `PeerList` type to be parametrized by
the type of values associated with a `PeerId`. This enables
easier testing of the peer list without having to construct
light client`Instance`s.

This commit additionally introduces a safer API by taking into
account the invariant associated with a `PeerList`, which lets
us return unconditionally the value associated with the primary peer.

In turn, this simplifies the verification loop in the supervisor,
which is now expressed more simply as a recursive function whose
correctness is more obvious than the imperative one, at least to me.