light-client: improve supervisor ergonomics (#403)
* light-client: turn Handle into a trait

As we start to depend on the surface of the `Handle` we benefit from it
being a trait that can be implemented on a per need basis. This will
result in less overhead constructing object graphs in places where we
wannt to assert behaviour of other types in remote places, i.e. the
light-node rpc server. Overall we hope for an increased ease in writing
tests on module level.

Ref #219

* Add changes entry

* Remove async trait methods

* Make supervisor more CSP

* Improve crate imports

* Relax mutability on Handle

* Provide noop default implementations

* Update adr status

* Handle crossbeam errors honestly

* Add changes entry

* Mention ADR in change entry

* Correct PR link

* Simplify example