Look for trusted or verified blocks in bisection loop
Only looking for blocks with `Status::Verified` might miss
a block that has previously gone through fork detection and
marked as `Trusted`. We would then fetch it again from the
peer, go through verification again, and then override its
status to `Verified`, thus downgrading it from `Trusted` to
`Verified`.

This commit fixes this by asking the light store for blocks which
are either trusted or verified.