After a bit of afterthought and as a result of @Shivani912 comment, decided to make `round` an explicit parameter when constructing the commit.
The reason for this is that we want to allow generating votes only from
the header, but without explicitely given round, the generated votes
will point to the default 1. Setting the `round` afterwards in the Commit
struct will have no effect on those votes.

That's why `round` is now an explicit parameter, and there are two constructors:
`new()`, and `new_with_votes()`.