Refactor subscription mechanism to simplify
The previous iteration of the subscription mechanism relied on two
asynchronous tasks to manage subscriptions and route events from the
WebSockets connection to their relevant subscribers.

This version only uses one async task (the WebSocket driver, which must
interact asynchronously with the connection).

In addition to this, I tried experimenting with using generics instead
of trait objects for the transport layer for the `Client`. This works,
but the downside of using generics is that, if you want generalizability
at layers higher up, it comes at the cost of decreased readability of
the code.

Signed-off-by: Thane Thomson <connect@thanethomson.com>