Refactor interface entirely
After much consideration, I decided to do away with the `Transport` and
related traits entirely. It makes more sense, and results in better
ability to test, if we abstract the `Client` interface instead. This
also allows for greater flexibility when implementing different kinds of
transports, while still allowing for lazy/optional instantiation of the
subscription mechanism.

Additional integration tests are included here, as well as additional
documentation.

Signed-off-by: Thane Thomson <connect@thanethomson.com>