Refactor interface for subscription client
The interface now no longer provides an option to subscribe using a
specific buffer size and we assume unbounded buffers everywhere. This
may or may not be a safe assumption going forward and we'll need to
re-evaluate in future.

As such, I could remove all code relating to bounded buffers.

Additionally, this commit adds something of an integration test for the
WebSocket-based subscription client (with a full WebSocket server,
effectively).

Signed-off-by: Thane Thomson <connect@thanethomson.com>