Simplify RPC client feature config
This simplifies the available features for the RPC client to just two:

* `http-client`, which provides an HTTP-based RPC client that does not
  provide event subscription functionality.
* `websocket-client`, which provides a WebSocket-based RPC client that
  exclusively provides event subscription functionality.

By including either feature, the mock client functionality is included.

Originally we thought we could call the features simply `http` and
`websocket`, but Rust doesn't allow crate features' names to be the same
as any of the dependencies' names, and in this case we're making use of
the `http` crate as a dependency.

Signed-off-by: Thane Thomson <connect@thanethomson.com>