Refactor to remove ChannelTx mutability
ChannelTx::send now no longer needs to be mutable because the underlying
channel has no mutability requirement. This leads to viral changes
throughout the interfaces and clients (positive ones).

ChannelTx::send also doesn't need to be async, because it relies on an
unbounded channel. This also has some viral implications for other
methods throughout the RPC client. If, in future, we want to support
bounded channels, then we can consider making it async again. But we
probably shouldn't make it async if we don't need it to be.

Signed-off-by: Thane Thomson <connect@thanethomson.com>