Make timestamp serialization safe
Without this change, serde of timestamps would result in a panic if the
stamp was before the UNIX epoch. In discussion with @thanethomson, we
agreed this opened up a potential attack vector, since a node could be
sent messages causing it to crash on deserialization.

It also violated the documented contract implied by the `From` trait:

> Note: This trait must not fail. If the conversion can fail, use TryFrom.

https://doc.rust-lang.org/std/convert/trait.From.html

The fix is simply to use `impl From(DateTime<Utc>) for SystemTime`
supplied by chrono.