Add tendermint-rpc binary
This is unfortunately a big commit. It primarily adds an optional binary
application (behind the "cli" feature flag) that allows you to easily
perform queries against a Tendermint endpoint via several protocols
(http, https, ws, wss).

It only contains support at present for Client queries, and
SubscriptionClient queries (i.e. subscriptions) are planned for
subsequent commits.

Some of the bulk of this commit is due to the fact that we need to be
able to parse queries (e.g. from tx_search), so I added a PEG-based
parser for queries along with some tests.

Signed-off-by: Thane Thomson <connect@thanethomson.com>