Refactor Light Client to remove trait objects
This commit refactors most of the public API of the Light Client to
remove any and all use of embedded trait objects. In order to accomplish
this, I also had to restructure the `VerificationPredicates` as
functions (it's not clear to me right now as to where anyone would
override these predicates).

Signed-off-by: Thane Thomson <connect@thanethomson.com>