p2p: replace `prost-amino` with `prost`
The only message that `prost-amino` was still used for was
`AuthSigMessage`, and the only way an amino-encoded version of that
message differs from a proto is the amino type prefix on the `pub_key`
field, i.e. `tendermint/PubKeyEd25519`.

This commit adds a test vector for a serialized amino message, and then
changes `AuthSigMessage` to use `prost-derive`:

- An `AuthSigMessage::new` function takes care of adding the Amino
  prefix to the public key
- A `TryFrom` impl for `proto::p2p::AuthSigMessage` takes care of
  verifying and removing it

With that, `prost-amino` is no longer necessary and this commit
therefore removes the last dependency on it.