light-client: Replace Io trait with async variant
While invasive this change transforms the light-client to be async all
the way. This touches all places where some I/O is taking place.
Achieved by roping in the async-trait[0] crate and dropping the existing
`block_on` implementation. Lo, this drops the timeout which guarded the
side-effectful operation, which will be reintroduced in a follow-up
change.

[0] https://github.com/dtolnay/async-trait

Signed-off-by: xla <self@xla.is>