Update to deprecations in ed25519 1.3
The most recent release of ed25519 API deprecates the standalone
SIGNATURE_LENGTH constant in favor of the Signature::BYTE_SIZE
associated item. Also, the Signature::new constructor is deprecated
in favor of Signature::from_bytes.
Update the code to not trigger the deprecation lints.

In turn, deprecate what was formerly the reexport of
that constant in the tendermint API.