Add domain types (updated) and fix integration tests for Tendermint v0.35.0 (#1022)
* tendermint: add From<Infallible> for Error

* Use hex for AppHash Debug formatting

* Regenerate protobuf types using tools/proto-compiler.

* Regenerate protos against tendermint v0.35.0-rc3.

* Remove SetOption-related code in tendermint-abci.

* Update imports to reflect protobuf movements.

The BlockParams and ConsensusParams structs moved out of the ABCI
protos.

* Update tendermint-abci to reflect upstream proto changes.

* p2p: treat all non-Ed25519 keys as unsupported

This fixes a compile error introduced by upstream proto changes that add
an Sr25519 variant.

* Improve ABCI response code modeling with NonZeroU32

The previous data modeling allowed a user to construct an `Err(0)` value that
would be serialized and deserialized as `Ok`.

* Use the Bytes type in ABCI protos.

* Add conversions between protobuf and chrono types.

* tendermint: define Serde for Block in terms of RawBlock

This changes the Serialize/Deserialize implementations for Block to convert
to/from the proto-generated `RawBlock`, and use the derived serialization for
that type.

This is much cleaner and more maintainable than keeping Serde annotations for
each sub-member of the data structure, because all of the serialization code is
kept in one place, and there's only one validation path (the TryFrom conversion
from RawBlock to Block) that applies to both kinds of serialization.

* tendermint: simpler transaction modeling in Block

This changes the Block type to hold the transactions as a plain
`Vec<Vec<u8>>`, rather than as a custom `abci::transaction::Data` type
(which is itself a wrapper for an `Option<Vec<Transaction>>`, the
`Transaction` type being a wrapper for a `Vec<u8>`).

This also updates the `Block` getter functions; it's not clear (to me)
why they're there, since they access the public fields and aren't used
anywhere else.

* rpc: take over tendermint::abci

The existing contents of the `tendermint::abci` module note that they're
only for the purpose of implementing the RPC endpoints, not a general
ABCI implementation.

Moving that code from the `tendermint` crate to the `tendermint-rpc`
crate decouples the RPC interface from improvements to the ABCI domain
modeling. Eventually, it would be good to eliminate this code and align
it with the new ABCI domain types.

* tendermint: add ABCI domain types.

These types mirror the generated types in tendermint_proto, but have better
naming.

The documentation for each request type is adapted from the ABCI Methods and
Types spec document. However, the same logical request may appear three
times, as a struct with the request data, as a Request variant, and as a
CategoryRequest variant.

To avoid duplication, this PR uses the `#[doc = include_str!(...)]`
functionality stabilized in Rust 1.54 to keep common definitions of the
documentation.

* tendermint: eliminate &'static str errors in ABCI domain types.

* Merge `abci::params::ConsensusParams` with `consensus::Params`.

The code in the `abci` module had more complete documentation from the ABCI
docs, so I copied it onto the existing structure.

* Add hex encoding Serde attribute to Sr25519 keys.

* Replace integers with `block::Height`, `vote::Power`

* Replace integer with block::Round

* Fix tools build by using correct imports

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Fix clippy complaints in tools

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Fix clippy warning

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Fix clippy lints

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Fix more clippy lints

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Fix deprecation notices from ed25519 crate

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Regenerate protos for Tendermint v0.35.0

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Fix raw bytes conversion in tests/docs

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Add Tendermint v0.35.0 Docker config

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Add Tendermint v0.35.0 Docker image for ABCI integration testing

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Remove RPC deserialization tests

The support files for these tests were manually generated some time ago.
We should rather favour testing with the `kvstore_fixtures` tests, whose
fixtures are automatically generated by our rpc-probe tool.

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Reformat Docker folder readme

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Bump version of Tendermint used in ABCI integration test

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Bump version of Tendermint used in rpc probe

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Bump version of Tendermint used in kvstore integration test

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Bump version of Tendermint used in proto-compiler

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Regenerate kvstore fixtures with rpc-probe for Tendermint v0.35.0

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Update kvstore integration test to accommodate newly generated fixtures

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Update RPC tests and data structures to accommodate Tendermint v0.35.0 changes

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Update ABCI encoding scheme to accommodate breaking change in https://github.com/tendermint/tendermint/issues/5783

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Bump Tendermint version in GitHub Actions kvstore integration test

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Add changelog entries to capture breaking changes

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Change tx hash encoding from base64 to hex and update tests

Signed-off-by: Thane Thomson <connect@thanethomson.com>

* Add changelog entry for /tx endpoint change

Signed-off-by: Thane Thomson <connect@thanethomson.com>

Co-authored-by: Henry de Valence <hdevalence@penumbra.zone>
Co-authored-by: Henry de Valence <hdevalence@hdevalence.ca>