p2p: Rework transport to be async (#1116)
* p2p: Update cargo metadata

Signed-off-by: xla <self@xla.is>

* p2p: Rework transport to be async

While it was well-intended the reality is that it's going to be
impractical and hinder potential adoption trying to not give in to the
sweet siren sounds of the async ecosystem. The author since has changed
their stance and adopted a conviction that well tuned and guard-railed
async can be workable and lead to maintainable solutions. This
change-set is in preparation of bringing in peer and supervisor
abstractions as described in the ADR.

Signed-off-by: xla <self@xla.is>