rpc: Eliminate abci::Gas
In tendermint-rpc, the only place where Gas newtype is still used
is in the response to /broadcast_tx_commit, which is a development-only
endpoint. Leave the response struct defined through tx_commit::TxResult
for the time being.
Define serialization of abci::response::CheckTx (currently not used)
to match that of DeliverTx.